fix(ihe): get metriport content from doc content
Ref. metriport/metriport-internal#1350
