+++ packages/api/src/command/medical/document/document-reconvert.ts
@@ -2,6 +2,7 @@ import { DocumentReference } from "@medplum/fhirtypes";
 import { resourceTypeForConsolidation } from "@metriport/api-sdk";
 import { S3Utils } from "@metriport/core/external/aws/s3";
 import { executeAsynchronously } from "@metriport/core/util/concurrency";
+import { getMetriportContent } from "@metriport/core/external/fhir/shared/extensions/metriport";
 import { out } from "@metriport/core/util/log";
 import { isMimeTypeXML } from "@metriport/core/util/mime";
 import { capture } from "@metriport/core/util/notifications";
@@ -14,7 +15,6 @@ import { convertCDAToFHIR } from "../../../external/fhir-converter/converter";
 import { getDocumentsFromFHIR as getDocumentsFromFHIRServer } from "../../../external/fhir/document/get-documents";
 import { countResources } from "../../../external/fhir/patient/count-resources";
 import { downloadedFromHIEs } from "@metriport/core/external/fhir/shared/index";
-import { getMetriportContent } from "../../../external/fhir/shared/extensions/metriport";
 import { Config } from "../../../shared/config";
 import { errorToString } from "../../../shared/log";
 import { formatNumber } from "@metriport/shared/common/numbers";
