+++ packages/api/src/external/fhir/document/index.ts
@@ -24,6 +24,8 @@ import {
   HumanName as CWHumanName,
 } from "@metriport/commonwell-sdk";
 import { Gender } from "@metriport/commonwell-sdk/src/models/demographics";
+import { metriportDataSourceExtension } from "@metriport/core/external/fhir/shared/extensions/metriport";
+import { cwExtension } from "@metriport/core/external/commonwell/extension";
 import dayjs from "dayjs";
 import isToday from "dayjs/plugin/isToday";
 import { sortBy, uniqBy } from "lodash";
@@ -32,8 +34,6 @@ import MetriportError from "../../../errors/metriport-error";
 import { capture } from "../../../shared/notifications";
 import { Util } from "../../../shared/util";
 import { CWDocumentWithMetriportData } from "../../commonwell/document/shared";
-import { cwExtension } from "../../commonwell/extension";
-import { metriportDataSourceExtension } from "../shared/extensions/metriport";
 import { toFHIRSubject } from "@metriport/core/external/fhir/patient/index";
 dayjs.extend(isToday);
