+++ packages/core/src/external/carequality/extension.ts
@@ -1,9 +1,9 @@
-import { Extension } from "@medplum/fhirtypes";
+import { Extension, DocumentReferenceContent } from "@medplum/fhirtypes";
 import { MedicalDataSource } from "@metriport/api-sdk";
 import { dataSourceExtensionDefaults } from "../fhir/shared/extensions/extension";
-import { DataSourceExtension } from "../fhir/shared/extensions/metriport";
+import { MetriportDataSourceExtension } from "../fhir/shared/extensions/metriport";
 
-export const cqExtension: DataSourceExtension = {
+export const cqExtension: MetriportDataSourceExtension = {
   ...dataSourceExtensionDefaults,
   valueCoding: {
     ...dataSourceExtensionDefaults.valueCoding,
@@ -13,4 +13,8 @@ export const cqExtension: DataSourceExtension = {
 
 export function isCarequalityExtension(e: Extension): boolean {
   return e.valueCoding?.code === cqExtension.valueCoding.code;
+}
+
+export function isCarequalityContent(content: DocumentReferenceContent): boolean {
+  return content.extension?.some(isCarequalityExtension) === true;
 }
